feat: /acapela today command in slack
